move unique tasks outside of Behaviour.hpp
